requirements.txt
asgiref==3.3.0
Django==3.1.3
django-autoslug==1.9.8
Pillow==8.0.1
python-slugify==4.0.1
pytz==2020.4
sqlparse==0.4.1
text-unidecode==1.3
